+++ src/hackerspace_online/settings.py
@@ -21,6 +21,7 @@ env = environ.Env(
     ALLOWED_HOSTS=(list)
 )
 
+
 project_root = environ.Path(__file__) - 3  # "/"
 PROJECT_ROOT = project_root()
 BASE_DIR = project_root('src')  # "/src/"
@@ -80,8 +81,6 @@ SHARED_APPS = (
 )
 
 TENANT_APPS = (
-    'django.contrib.contenttypes',
-
     'django.contrib.auth',
     'django.contrib.sessions',
     'django.contrib.messages',
@@ -181,7 +180,7 @@ INSTALLED_APPS = (
     'hackerspace_online',
 
     # django storages
-    'storages', 
+    'storages',
 
     # local apps
     'quest_manager',
@@ -212,6 +211,48 @@ MIDDLEWARE = [
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
     'django.middleware.security.SecurityMiddleware',
 ]
+
+DB_LOGS_ENABLED = env('DB_LOGS_ENABLED', default=False)
+
+if DB_LOGS_ENABLED:
+    MIDDLEWARE.insert(0, 'hackerspace_online.middleware.ForceDebugCursorMiddleware')
+    print(MIDDLEWARE)
+    LOGS_PATH = os.path.join(os.sep, 'tmp', 'bytedeck')
+
+    if not os.path.exists(LOGS_PATH):
+        os.mkdir(LOGS_PATH)
+
+    LOGGING = {
+        'version': 1,
+        'disable_existing_loggers': False,
+        'formatters': {
+            'verbose': {
+                'format': '{levelname} | {asctime} | {process:d} {thread:d} | {message}',
+                'style': '{',
+            },
+        },
+        'handlers': {
+            'console': {
+                'level': 'DEBUG',
+                'class': 'logging.StreamHandler',
+                'formatter': 'verbose',
+            },
+            'file': {
+                'level': 'DEBUG',
+                'class': 'logging.handlers.TimedRotatingFileHandler',
+                'when': 'M',  # Every minute (so that it would be much easier to view queries rather than by hour or day)
+                'filename': os.path.join(LOGS_PATH, 'queries.log'),
+                'formatter': 'verbose',
+            }
+        },
+        'loggers': {
+            'django.db.backends': {
+                'handlers': ['file'],
+                'level': 'DEBUG',
+            },
+        }
+    }
+
 
 ROOT_URLCONF = 'hackerspace_online.urls'
 
@@ -345,7 +386,7 @@ if USE_S3:
     # AWS settings
     AWS_ACCESS_KEY_ID = env('AWS_ACCESS_KEY_ID')
     AWS_SECRET_ACCESS_KEY = env('AWS_SECRET_ACCESS_KEY')
-    
+
     # S3 settings
     AWS_STORAGE_BUCKET_NAME = env('AWS_STORAGE_BUCKET_NAME')
     AWS_S3_CUSTOM_DOMAIN = env('CDN_static')
@@ -353,12 +394,12 @@ if USE_S3:
         "ACL": "public-read",
         "CacheControl": "max-age=86400"
     }
-    
+
     # S3 Static Files
     STATICFILES_LOCATION = 'static'
     STATIC_URL = f'https://{AWS_S3_CUSTOM_DOMAIN}/{STATICFILES_LOCATION}/'
     STATICFILES_STORAGE = 'storage.custom_storages.StaticStorage'
-    
+
     # Media Files
 
     # S3 public media files
@@ -370,12 +411,12 @@ if USE_S3:
     # For any implementation in future, Refer https://testdriven.io/blog/storing-django-static-and-media-files-on-amazon-s3/
     PRIVATE_MEDIAFILES_LOCATION = 'private_media'
     PRIVATE_FILE_STORAGE = 'storage.custom_storages.PrivateMediaStorage'
-    
+
 else:
-    
+
     # The absolute path to the directory where `collectstatic` will move the static files to.
     STATIC_ROOT = env('STATIC_ROOT', default=os.path.join(PROJECT_ROOT, "_collected_static"))
-    
+
     # The absolute path to the directory where uploaded media files will be saved to
     MEDIA_ROOT = env('MEDIA_ROOT', default=os.path.join(PROJECT_ROOT, "_media_uploads"))